Fixed ternary operations (#2)
When writing a ternary expression in C#, the compiler generates MSIL byte code with 4 sections:

    Condition check statement. If true it branches to section 3, otherwise if falls into section 2
    False logic section. Will pop a final value onto the evaluation stack and branch to section 4
    True logic section. Will pop a final value onto the evaluation stack and fall into section 4
    Final execution section. Contains the logic that handles the single value on the evaluation stack and performs the final operation with it (e.g. assignment).

This is easily handled via runtime interpretation when storing values on an actual stack, but handling this when transpiling is tricky. We can't statically guarantee if section 2 or 3 will be entered, and thus the transpilation process will end up with both section's resulting expressions in our expression stack.

To work around this, we can infer that if a simple branch is encountered with an expression on the expression stack, that we are in a scenario like section 2 above. This means we have to write the expression in the expression stack to an assignment statement before we perform the branch in C.

Then when we get to the instruction that the branch targets, we know we should have a different expression in the expression stack. So before we write the statements for the target instruction, we write another statement to save the new expression in the stack to the checkpoint variable.

Then popping the checkpoint variable onto the stack, all code in section 4 of the ternary process uses the checkpoint variable. This allows the C code to correctly utilize the correct expression based on the correct ternary branch.