Native globals on transpile (#6)
Adds the ability to use attributes to mark a static field as being a natively declared global on transpile. This will cause the transpiler to replace any calls to get or set values for that field to the name of the native value provided, along with including any required headers for it.