Transpile to a single source file (#7)
In order to support some use cases (such as eBPF generation), it is useful to manually combine several transpiled headers and source files into a single source-only file.  

This commit updates the manifest to add a `SingleGeneratedSourceFileName` property, which specifies that all transpilation should occur into a single source file created in the output directory for these use cases.