Make globals first class citizens  (#8)
Static fields / globals were implemented as a hacky part of the pipeline as an extension of how instance fields were implemented. This is causing more and more problems as more customization is needed for them.

So this PR gives globals/static fields their own definition type, and allows it to run through the dependency graph and pipeline the same way methods and types are.