Moved NativeFunctionCall resolution to definition time
Previously, if you marked a method as being a native function call, we
would track the method as a dot net method being defined, and have
special code that recognized the desire for it to be a native function
when creating a `MethodConversionInfo`.

This commit simplifies things a bit by recognizing the method as having
this attribute before creating the definition for it. If we see it's
marked with a `NativeFunctionCallAttribute` then we instead form a
`NativeDefinedMethod` instead of a `DotNetMethod`.

This saves us from having to analyze the IL and special casing some code
paths.