Refactored native global attribute handling
When marking a static field with a `NativeGlobalAttribute`, the logic to
handle that was spread around in op code handlers.

To simplify this, we instead look for the attribute as part of adding
the type to the definition catalog. If we see a type with a static
field, and that static field has the `NativeGlobalAttribute` on it, then
we define that field as a `NativeDefinedGlobal` instead of a
`DotNetDefinedGlobal`.