Add support for transpiling unreferenced globals
Manifests can now specify globals to transpile that are not referenced by other .net code. The
main use case for this is specifying values stored in ELF binaries for BPF programs.

This also allows specifying initial values for globals. While this shouldn't need to be done
via an attribute, pulling literal value assignments out of static constructors isn't the most
trivial and is something to do later on.