Support generic methods defined in another assembly
When a generic method is called that is defined in a different assembly
than the calling code, the method call does not have the real generic
parameter names used at definition time (e.g. `T`, `U`, etc..).  They instead
use the parameters that denote the index of the generic parameter (e.g. !!2).

This was making the definitions and conversion info types not able to be found
by the calling code, and thus there was no way to know how to invoke the call.

This is fixed by normalizing generic identifiers to make sure they always use
the index based parameter naming.