Add base class for array defined types.
The base class for array defined types allows for easier access
to a generic way of getting the size of an array. The size of an
array will be different for different implementations (heap vs
constant sized as an example).

This base class also allows for easier access to the real type
of the array's element, instead of guessing.