Fixed bug with method calls that returned a pointer type.
For simplicity when we encounter a method call in the IL, we push that IL
onto the expression stack. This causes the generated c code for the function
call to be written where it's used instead of on it's own line.

However, this does not work when that function returns a pointer type, as we
can't guarantee that the function's usage won't be used in a way that would
violate C syntax rules. For example, a function that returns a pointer
to an int could end up looking like `(*func()) = 25;` for an assignment, which
is not valid c.

To fix this, if a method call would return a pointer, we store the value in a
temporary local. This ensures different scenarios work as expected.