Fix accidental pointer dereference on assignment
When a pointer value was being stored in a local that takes a pointer, it
was being incorrectly dereferenced causing compile errors. The code was updated
to not dereference a value in a store opcode.

Additionally, a bug was found where pointer parameters were being transpiled as
non-pointers, causing them to be incorrect.

This commit fixes both of these issues.