Fix native array support.
Native global arrays that are not declared with the StaticallySizedArrayAttribute
should be treated as statically sized arrays with the bounds checking disabled. This
is needed because we can't assume a natively defined array in C has a specific heap allocated
structure, nor do we know the size without additional attributes.