Fixed bug with ++ operator.
When a ++ operator is used within an array index operation, and a static field
index variable is what is being incremented, the MSIL relies on loading the index
variable onto the evaluation stack, duplicating it, then incrementing the static field
*before* performing the index. This means that when we perform the index it's after
the increment, and thus the transpiled code reads the wrong array value.

This is fixed by using the same variable replacement trick in the expression stack's
syntax tree. By saving off the value of the index static field into a temporary local
before incrementing, and using the temporary local for the index operation, we get the
desired behavior.

Added comment