Add CI validation for prs (#16)
Adds a github action to validate that all samples have been built and no
changes to sample project generated code is missing from the commit.

This helps make it noticable in code reviews if a PR causes a change in
generation behavior (potentially catching regressions), or at least pin
point when a specific code generation change was made.