Fixed double asterisk on string parameters.
With the addition of reference type support, string parameters are being counted as
pass by reference when they weren't before (without an explicit ref). This meant that
when trying to form the method declaration we would prepend an asterisk to the parameter.

This breaks because the type for strings is already `char*` and thus you'd end up with
`char* *someString`, which isn't what's desired most of the time.

This commit then ensures that if the parameter type's native name already ends in an asterisk
(e.g. is an explicit pointer) then it will not have duplicate asterisks.

This is probably going to be buggy, and will probably need to be revisited later.