Fixed call syntax for non-virtual class methods.
When a non-virtual class method is called, the IL will sometimes use the
`callvirt` MSIL opcode for it.  This was causing the method call expression
to be flagged as a virtual call, which would have used the syntax that expects
the type to have a function pointer on the struct. Since the function isn't
virtual we don't generate a function pointer field for it, and thus it fails
to compile.

This fixes this by ensuring that method calls are only marked as virtual if the
method they are calling is actually virtual.