Add function to untrack dotnet reference types (#24)
Adds a new function that handles untracking an object from the garbage
collection algorithm currently in place. For the initial implementation
this consists of decrementing the reference count and calling `free()`
if the reference count drops to zero.

The function nulls out the caller's pointer as well, to prevent and
discourage native C callers from calling it after its been freed.

This does not contain functionality to untrack nested reference types.