Prevent track+untrack in returns.
When we return a .net reference type, we want to make sure it is
marked as tracked so it does not get freed as part of returning
from a function. Likewise, when we return from a transpiled function
we want to mark each local reference type as untracked so they do
not leak.

However, if the reference type we are returning is a local variable
then this causes us to first track then untrack. This is extra work
that isn't needed, and thus we can detect this and leave it out.