Untracking a reference type should recursively prep reference type (#25)
When calling `UnTrack()` on a .net reference type, before we deallocate
it we need to call `UnTrack()` on any reference type that the original
untracked type has fields for.

In order to accomplish this, the `ReferenceTypeBase` type needs a
pointer to the `PrepForFree()` function for the specific reference type
that the `ReferenceTypeBase` is attached to. This function pointer is
then invoked before deallocation to perform the correct preparations.