Stloc should track the stored reference type
Previously, I did not have stored reference types with a stloc op code
have the local tracked after assignment because the call handler was already
tracking the instance. However, this was the wrong call because it would cause
storing a 2nd value on top of the same local to be decremented too many times,
causing a use after free scenario.

This exposed another bug where if a reference type was used in the final expression
utilized by a return statement, then that reference type would be untracked and freed
prior to being invoked by the return statement, causing a use after free crash.

This was fixed by having functions explicitely store a return value local. This allows
us to compute the return value into a local before calling UnTrack() on all locals.
This accurately avoided untracking a reference type after free-ing it.