Replaced IsPointer expression values with pointer depth
In order to support reference types with the `ref` tag on it, we need
to support double pointers. Therefore, a singular `IsPointer` flag wasn't
cutting it.

This replaced the expression logic to think about pointers as pointer depth
instead of a single boolean.

This also replaced the `AddressOf` and `Dereference` expressions with one that
adjusts the expression to a specified target pointer depth (up or down). This
simplifies some of the logic for ensuring two expressions are equal.