Fixed bugs with statically sized arrays
A regression was found with statically sized arrays of structs, in that
the ldelem op code was improperly giving the ArrayIndexExpression a
resulting type of the array instead of the array's elements. This would
then fail with an ldfld opcode, because the array doesn't contain that
field. This fixes it by making sure that ldelem always gives the
ArrayIndexExpression the resulting type of the element.

Another bug was found with the ldlen op code where it was hard coded to
check for length via a `length` field. This obviously is incorrect for
statically sized arrays, so this was fixed to get the array size via the
`GetArraySizeExpression` method call.