Generate create function for ref types that aren't instantiated
When a .net reference type is used by a function (e.g. provided as an
argument) but that reference type is never created by transpiled C#
code, then the `__Create()` function for that type was never being
generated.

This is a bug because the `__Create()` function is needed if any C code
is going to generate an instance for the transpiled C# code to use, and
without it the necessary function pointers and reference tracking
mechanisms won't be created.

This fixes that bug and forces `__Create()` functions to be generated
for any .net ref type that is referenced by transpiled methods/types.